Test motif enrichment using a background set of sequences

Test for motif enrichment within a set of sequences using a background set to derive a NULL hypothesis

testMotifEnrich(
  pwm,
  stringset,
  bg,
  var = "iteration",
  model = c("poisson", "iteration", "quasipoisson", "negbinom"),
  sort_by = c("p", "none"),
  mc.cores = 1,
  ...
)

Arguments

pwm

A Position Weight Matrix or list of PWMs

stringset

An XStringSet with equal sequence widths

bg

An XStringSet with the same sequence widths as the test XStringset

var

A column in the mcols element of bg, usually denoting an iteration number

model

The model used for analysis

sort_by

Column to sort results by

mc.cores

Passed to mclapply

...

Passed internally to countPWM

Value

A data.frame with columns: `sequences`, `matches`, `expected`, `enrichment`, `Z` and `p`, with additional columns `est_bg_rate` (Poisson) or `sd_bg`, `n_iter` and `iter_p` (Iterations). The numbers of sequences and matches refer to the test set of sequences, whilst expected is the expected number of matches under the Poisson or iterative null distribution. The ratio of matches to expected is given as `enrichment`, along with the Z score an p-value. The Z score is only representative under the Poisson model, but is used to estimate p-values under the iterative approach. Under this latter approach, the sd of the matches found in the background sequences is also given, along with the number of iterations and the p-values from permutations testing the one-sided hypothesis hypothesis for enrichment.

It may also be worth noting that if producing background sequences using makeRMRanges with `replace = TRUE` and `force_ol = TRUE`, the iterative model corresponds to a bootstrap, given that the test sequences will overlap the background sequences and background ranges are able to be sampled with replacement.

Details

This function offers two alternatives for assessing the enrichment of a motif within a set of sequences, in comparison to a background set of sequences. Enrichment using an iterative model assumes that the background sequences can be subset using a column in their `mcols()` element and these are iterated through to derive a mean and sd operating under the null hypothesis. These values are then used in a Z test to provide an estimate of motif enrichment. Whilst relying on few distributional assumptions and relying on the Central Limit Theorem for the Z score, this can be a time consuming and computationally demanding task.

Testing on real datasets has revealed the the background set of counts commonly follows a Poisson distribution across multiple iterations. As a less time-consuming and computationally-demanding approach, is also offered as a testing strategy. No iterations are required, but a background set of sequences which is far greater than the sequences of interest is advised, with background sets >1000 times larger being preferred.

Examples

## Load the example peaks & the sequences
data("ar_er_peaks")
data("ar_er_seq")
sq <- seqinfo(ar_er_peaks)
## Now sample size-matched ranges 10 times larger. In real-world analyses,
## this set should be sampled as at least 1000x larger, ensuring features
## are matched to your requirements. This example masks regions with known N
## content, including centromeres & telomeres
data("hg19_mask")
set.seed(305)
bg_ranges <- makeRMRanges(
  ar_er_peaks, GRanges(sq)[1], exclude = hg19_mask, n_iter = 10
)

## Convert ranges to DNAStringSets
library(BSgenome.Hsapiens.UCSC.hg19)
#> Loading required package: BSgenome
#> Loading required package: BiocIO
#> Loading required package: rtracklayer
#> 
#> Attaching package: ‘rtracklayer’
#> The following object is masked from ‘package:BiocIO’:
#> 
#>     FileForFormat
genome <- BSgenome.Hsapiens.UCSC.hg19
bg_seq <- getSeq(genome, bg_ranges)

## Test for enrichment of the ESR1 motif
data("ex_pwm")
esr1 <- ex_pwm$ESR1
testMotifEnrich(esr1, ar_er_seq, bg_seq, model = "poisson")
#>   sequences matches expected enrichment        Z            p          fdr
#> 1       229      62     12.3    5.04065 14.17111 6.735011e-24 6.735011e-24
#>   est_bg_rate
#> 1  0.05371179

## Test all motifs
testMotifEnrich(ex_pwm, ar_er_seq, bg_seq, model = "poisson")
#>       sequences matches expected enrichment          Z            p
#> FOXA1       229     334    178.4  1.8721973 11.6496310 2.693669e-25
#> ESR1        229      62     12.3  5.0406504 14.1711088 6.735011e-24
#> ZN281       229      63     48.9  1.2883436  2.0163443 5.287479e-02
#> ZN143       229       1      0.6  1.6666667  0.5163978 4.511884e-01
#> ANDR        229      61     64.9  0.9399076 -0.4841080 7.092503e-01
#>                fdr est_bg_rate
#> FOXA1 1.346835e-24 0.779039301
#> ESR1  1.683753e-23 0.053711790
#> ZN281 8.812465e-02 0.213537118
#> ZN143 5.639855e-01 0.002620087
#> ANDR  7.092503e-01 0.283406114