Find all PWM matches within an XStringSet

Find all PWM matches within a set of sequences

getPwmMatches(
  pwm,
  stringset,
  rc = TRUE,
  min_score = "80%",
  best_only = FALSE,
  break_ties = c("all", "random", "first", "last", "central"),
  mc.cores = 1,
  ...
)

Arguments

pwm

A Position Weight Matrix, list of PWMs or universalmotif list

stringset

An XStringSet

rc

logical(1) Also find matches using the reverse complement of pwm

min_score

The minimum score to return a match

best_only

logical(1) Only return the best match

break_ties

Method for breaking ties when only returning the best match Ignored when all matches are returned (the default)

mc.cores

Passed to mclapply if passing multiple PWMs

...

Passed to matchPWM

Value

A DataFrame with columns: seq, score, direction, start, end, from_centre, seq_width, and match

The first three columns describe the sequence with matches, the score of the match and whether the match was found using the forward or reverse PWM. The columns start, end and width describe the where the match was found in the sequence, whilst from_centre defines the distance between the centre of the match and the centre of the sequence being queried. The final column contains the matching fragment of the sequence as an XStringSet.

When passing a list of PWMs, a list of the above DataFrames will be returned.

Details

Taking a set of sequences as an XStringSet, find all matches above the supplied score (i.e. threshold) for a single Position Weight Matrix (PWM), generally representing a transcription factor binding motif. By default, matches are performed using the PWM as provided and the reverse complement, however this can easily be disabled by setting rc = FALSE.

The function relies heavily on matchPWM and Views for speed.

When choosing to return the best match (best_only = TRUE), only the match with the highest score is returned for each sequence. Should there be tied scores, the best match can be chosen as either the first, last, most central, all tied matches, or choosing one at random (the default).

Examples

## Load the example PWM
data("ex_pwm")
esr1 <- ex_pwm$ESR1

## Load the example Peaks
data("ar_er_seq")

## Return all matches
getPwmMatches(esr1, ar_er_seq)
#> DataFrame with 6 rows and 8 columns
#>         seq     score direction     start       end from_centre seq_width
#>   <integer> <numeric>  <factor> <integer> <integer>   <numeric> <integer>
#> 1        23   19.7743         R       151       165         -42       400
#> 2        26   18.7071         F       303       317         110       400
#> 3        38   18.0880         R       326       340         133       400
#> 4       131   19.0738         F       223       237          30       400
#> 5       140   18.4739         F       232       246          39       400
#> 6       166   17.5586         R       209       223          16       400
#>             match
#>    <DNAStringSet>
#> 1 AGGTCAGCCTGCCCT
#> 2 AGGACACCGTGACCT
#> 3 AGGTCACCCTGGCCC
#> 4 GGGTCAGCATGACCT
#> 5 AGGACACACTGACCT
#> 6 AGGTTAGCCTGACCT

## Just the best match
getPwmMatches(esr1, ar_er_seq, best_only = TRUE)
#> DataFrame with 6 rows and 8 columns
#>         seq     score direction     start       end from_centre seq_width
#>   <integer> <numeric>  <factor> <integer> <integer>   <numeric> <integer>
#> 1        23   19.7743         R       151       165         -42       400
#> 2        26   18.7071         F       303       317         110       400
#> 3        38   18.0880         R       326       340         133       400
#> 4       131   19.0738         F       223       237          30       400
#> 5       140   18.4739         F       232       246          39       400
#> 6       166   17.5586         R       209       223          16       400
#>             match
#>    <DNAStringSet>
#> 1 AGGTCAGCCTGCCCT
#> 2 AGGACACCGTGACCT
#> 3 AGGTCACCCTGGCCC
#> 4 GGGTCAGCATGACCT
#> 5 AGGACACACTGACCT
#> 6 AGGTTAGCCTGACCT

## Apply multiple PWMs as a list
getPwmMatches(ex_pwm, ar_er_seq, best_only = TRUE)
#> $ESR1
#> DataFrame with 6 rows and 8 columns
#>         seq     score direction     start       end from_centre seq_width
#>   <integer> <numeric>  <factor> <integer> <integer>   <numeric> <integer>
#> 1        23   19.7743         R       151       165         -42       400
#> 2        26   18.7071         F       303       317         110       400
#> 3        38   18.0880         R       326       340         133       400
#> 4       131   19.0738         F       223       237          30       400
#> 5       140   18.4739         F       232       246          39       400
#> 6       166   17.5586         R       209       223          16       400
#>             match
#>    <DNAStringSet>
#> 1 AGGTCAGCCTGCCCT
#> 2 AGGACACCGTGACCT
#> 3 AGGTCACCCTGGCCC
#> 4 GGGTCAGCATGACCT
#> 5 AGGACACACTGACCT
#> 6 AGGTTAGCCTGACCT
#> 
#> $ANDR
#> DataFrame with 1 row and 8 columns
#>         seq     score direction     start       end from_centre seq_width
#>   <integer> <numeric>  <factor> <integer> <integer>   <numeric> <integer>
#> 1       132   18.9718         R       264       281        72.5       400
#>                match
#>       <DNAStringSet>
#> 1 TGTTCAGCTTTGTTTGCT
#> 
#> $FOXA1
#> DataFrame with 40 rows and 8 columns
#>           seq     score direction     start       end from_centre seq_width
#>     <integer> <numeric>  <factor> <integer> <integer>   <numeric> <integer>
#> 1           6   14.0940         R       187       198        -7.5       400
#> 2           8   14.4605         F       169       180       -25.5       400
#> 3          13   14.9185         F       179       190       -15.5       400
#> 4          32   14.2796         F       230       241        35.5       400
#> 5          69   15.3767         R       372       383       177.5       400
#> ...       ...       ...       ...       ...       ...         ...       ...
#> 36        201   14.2366         R       184       195       -10.5       400
#> 37        202   14.0131         R       236       247        41.5       400
#> 38        204   15.3697         R       349       360       154.5       400
#> 39        208   14.4605         R       274       285        79.5       400
#> 40        223   14.1175         R        63        74      -131.5       400
#>              match
#>     <DNAStringSet>
#> 1     TGTTTGCATTGC
#> 2     TGTTTACTTTCC
#> 3     TGTTTACACAGT
#> 4     TATTTACATAGC
#> 5     TGTTTACCTTGG
#> ...            ...
#> 36    TGTTTACTTAAA
#> 37    TGTTGACTTTGT
#> 38    TGTTTACTTTTC
#> 39    TGTTTACTTTCC
#> 40    TGTTTATTTTAG
#> 
#> $ZN143
#> DataFrame with 0 rows and 8 columns
#> 
#> $ZN281
#> DataFrame with 3 rows and 8 columns
#>         seq     score direction     start       end from_centre seq_width
#>   <integer> <numeric>  <factor> <integer> <integer>   <numeric> <integer>
#> 1        26   19.3263         F        95       109         -98       400
#> 2        37   18.5467         R        84        98        -109       400
#> 3        72   20.5418         R       364       378         171       400
#>             match
#>    <DNAStringSet>
#> 1 GAATGGGGGAGGGGC
#> 2 GGATGGGGGAAGGGG
#> 3 GGTAGGGGGAGGGGA
#>